bbc/bbc/model/src/requirements.txt
python-pachyderm==6.1.0
scikit-learn==0.24.2
pandas==1.2.4
psycopg2-binary~=2.7.7
nltk~=3.6.2